requirements.txt
altair==5.5.0
asgiref==3.9.2
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-filter==25.1
djangorestframework==3.16.1
easyocr==1.7.2
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
imageio==2.37.0
imutils==0.5.4
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
Markdown==3.9
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.6.0
networkx==3.5
ninja==1.13.0
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.32.1
pyarrow==21.0.0
pyclipper==1.3.0.post6
pydeck==0.9.1
PyMatting==1.1.14
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rembg==2.0.67
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.9.30
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
